Use double for mathsize parsing in convertMathSizeIfNeeded
This change updates the mathsize parsing logic in MathMLElement.cpp to use
double instead of float, providing better precision and addressing a
FIXME in the code.

* Source/WebCore/mathml/MathMLElement.cpp:
(WebCore::convertMathSizeIfNeeded): Update unitlessValue to double and
use value.toDouble(). Remove related FIXME.

+++ Source/WebCore/mathml/MathMLElement.cpp
@@ -157,10 +157,9 @@ static String convertMathSizeIfNeeded(const AtomString& value)
         return "1.5em"_s;
 
     // FIXME: mathsize accepts any MathML length, including named spaces (see parseMathMLLength).
-    // FIXME: Might be better to use double than float.
     // FIXME: Might be better to use "shortest" numeric formatting instead of fixed width.
     bool ok = false;
-    float unitlessValue = value.toFloat(&ok);
+    double unitlessValue = value.toDouble(&ok);
     if (!ok)
         return value;
     return makeString(FormattedNumber::fixedWidth(unitlessValue * 100, 3), '%');
